+++ eppo_core/src/events.rs
@@ -1,9 +1,3 @@
-mod event;
-mod event_delivery;
-mod event_dispatcher;
-mod queued_event;
-mod vec_event_queue;
-
 use std::{collections::HashMap, sync::Arc};
 
 use serde::Serialize;